Extract device logs modal adapter from main-app.js
