+++ app/templates/base.html
@@ -310,6 +310,7 @@
     <script src="{{ request.url_for('static', path='js/features/device/device-capability-detector.js') }}?v={{ cache_version }}"></script>
     <script src="{{ request.url_for('static', path='js/features/clipboard/clipboard-controller.js') }}?v={{ cache_version }}"></script>
     <script src="{{ request.url_for('static', path='js/features/network/network-info-modal-adapter.js') }}?v={{ cache_version }}"></script>
+    <script src="{{ request.url_for('static', path='js/features/device/device-logs-modal-adapter.js') }}?v={{ cache_version }}"></script>
     <script src="{{ request.url_for('static', path='js/core/main-app.js') }}?v={{ cache_version }}"></script>
     <script src="{{ request.url_for('static', path='js/vendor/docx-preview.min.js') }}?v={{ cache_version }}"></script>
     <script src="{{ request.url_for('static', path='js/features/clipboard/clipboard-view-adapter.js') }}?v={{ cache_version }}"></script>
